refactor: expose retained identity through object seam
